fix: build:npm command
