+++ packages/code-du-travail-modeles/src/modeles/common/types/index.ts
@@ -1,3 +1,3 @@
-export * from "./formula";
+export { CategoryPro16 } from "../../conventions/16_transports_routiers/salary";
 export * from "./salary";
 export * from "./seniority";
