+++ packages/code-du-travail-modeles/src/modeles/conventions/44_industries_chimiques/indemnite-licenciement.yaml
@@ -149,6 +149,7 @@ contrat salarié . convention collective . industries chimiques . indemnité de 
     Article 11 de l’Accord du 15 janvier 1991 relatif à la politique de l'emploi (En vigueur non étendu): https://www.legifrance.gouv.fr/conv_coll/article/KALIARTI000005846081?idConteneur=KALICONT000005635613
     Article 10 de la Convention collective nationale des industries chimiques et connexes du 30 décembre 1952: https://www.legifrance.gouv.fr/conv_coll/article/KALIARTI000005846022?idConteneur=KALICONT000005635613&origin=list#KALIARTI000005846022
 
+
 contrat salarié . convention collective . industries chimiques . indemnité de licenciement . catégorie professionnelle . licenciement économique . age . résultat conventionnel 2:
   remplace: contrat salarié . indemnité de licenciement . résultat conventionnel
   applicable si:
